first version
first version bug fixes
